Fix Movie object response

+++ app/src/main/java/com/example/tmdb/data/discover/entities/responses/MovieDiscoverResponse.kt
@@ -44,7 +44,7 @@ data class MovieDiscoverResponse(
         @SerializedName("video")
         val video: Boolean?,
         @SerializedName("vote_average")
-        val voteAverage: Int?,
+        val voteAverage: Float?,
         @SerializedName("vote_count")
         val voteCount: Int?
     ) : Parcelable
